LayoutTests/fast/mediastream/RTCPeerConnection-setLocalDescription-offer-expected.txt

Test RTCPeerConnection.setLocalDescription called with an RTCSessionDescription of type 'offer'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.signalingState is 'stable'
PASS pc.localDescription is null
PASS pc.pendingLocalDescription is null
PASS pc.currentLocalDescription is null

PASS Got stream
*** Add audioTrack.
PASS pc.getTransceivers().length is 1
PASS audioTransceiver.mid is null

PASS Got firstOffer
PASS pc.signalingState is 'stable'
PASS audioTransceiver.mid is null

PASS firstOffer set as local description
FAIL pc.signalingState should be have-local-offer. Was stable.
FAIL pc.localDescription should be [object RTCSessionDescription]. Was null.
FAIL pc.pendingLocalDescription should be [object RTCSessionDescription]. Was null.
PASS pc.currentLocalDescription is null
FAIL Error caught in promise chain: TypeError: null is not an object (evaluating 'pc.localDescription.type')
PASS successfullyParsed is true

TEST COMPLETE
